# pkg/handler/list.go
package handler

import (
	"ToDoApp"
	"github.com/gin-gonic/gin"
	"net/http"
	"strconv"
)

// @Summary      todo List
// @Security 	ApiKeyAuth
// @Description  create todo list
// @Tags         lists
// @ID 			 create-list
// @Accept       json
// @Produce      json
// @Param       input body ToDoApp.TodoList true "list info"
// @Success      200  {integer}   integer  1
// @Failure      400  {object}  errorResponse
// @Failure      404  {object} errorResponse
// @Failure      500  {object} errorResponse
// @Failure default {object} errorResponse
// @Router       /auth/lists [post]
func (h *Handler) createList(c *gin.Context) {
	userId, err := getUserId(c)
	if err != nil {
		return
	}

	var input ToDoApp.TodoList
	if err := c.BindJSON(&input); err != nil {
		newErrorResponse(c, http.StatusBadRequest, err.Error())
		return
	}
	id, err := h.services.TodoList.Create(userId, input)
	if err != nil {
		newErrorResponse(c, http.StatusInternalServerError, err.Error())
		return
	}
	c.JSON(http.StatusOK, map[string]interface{}{
		"id": id,
	})

}

// @Summary      todo List
// @Security 	ApiKeyAuth
// @Description  get todo list
// @Tags         lists
// @ID 			 get-list
// @Accept       json
// @Produce      json
// @Param       input body ToDoApp.TodoList true "list info"
// @Success      200  {integer}   integer  1
// @Failure      400  {object}  errorResponse
// @Failure      404  {object} errorResponse
// @Failure      500  {object} errorResponse
// @Failure default {object} errorResponse
// @Router       /auth/lists [get]
type getAllListsResponse struct {
	Data []ToDoApp.TodoList `json:"data"`
}

func (h *Handler) getAllLists(c *gin.Context) {
	userId, err := getUserId(c)
	if err != nil {
		return
	}
	lists, err := h.services.TodoList.GetAll(userId)
	if err != nil {
		newErrorResponse(c, http.StatusInternalServerError, err.Error())
		return
	}
	c.JSON(http.StatusOK, getAllListsResponse{
		Data: lists,
	})
}

func (h *Handler) getListById(c *gin.Context) {
	userId, err := getUserId(c)
	if err != nil {
		return
	}
	id, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		newErrorResponse(c, http.StatusBadRequest, "invalid id param")
		return
	}
	list, err := h.services.TodoList.GetById(userId, id)
	if err != nil {
		newErrorResponse(c, http.StatusInternalServerError, err.Error())
		return
	}
	c.JSON(http.StatusOK, list)
}

func (h *Handler) updateList(c *gin.Context) {
	userId, err := getUserId(c)
	if err != nil {
		return
	}
	id, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		newErrorResponse(c, http.StatusBadRequest, "invalid id param")
		return
	}

	var input ToDoApp.UpdateListInput
	if err := c.BindJSON(&input); err != nil {
		newErrorResponse(c, http.StatusBadRequest, err.Error())
		return
	}
	if err := h.services.TodoList.Update(userId, id, input); err != nil {
		newErrorResponse(c, http.StatusInternalServerError, err.Error())
		return
	}
	c.JSON(http.StatusOK, statusResponse{"ok"})
}
func (h *Handler) deleteList(c *gin.Context) {
	userId, err := getUserId(c)
	if err != nil {
		return
	}
	id, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		newErrorResponse(c, http.StatusBadRequest, "invalid id param")
		return
	}
	err = h.services.TodoList.Delete(userId, id)
	if err != nil {
		newErrorResponse(c, http.StatusInternalServerError, err.Error())
		return
	}
	c.JSON(http.StatusOK, statusResponse{
		Status: "ok",
	})
}